requirements.txt
# requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
pyserial==3.5
minimalmodbus==2.1.1
RPi.GPIO==0.7.1  # Only needed on Raspberry Pi